client/requirements_38.txt
munch>=2.5.0
rich>=12.0.1
aiohttp>=3.6.2
cached-property>=1.5.2
setuptools_scm>=6.3.2
asyncclick==8.0.1.3
aiofiles>=0.8.0
aiosqlite>=0.17.0
sysv_ipc>=1.1.0
requests>=2.26.0
python-dateutil>=2.8.2
anyio>=3.4.0
pyyaml>=5.4.1
pydantic>=1.9.0
treelib>=1.6.1
